Fixed bug with empty startdate
Using form.html without entering anything in the startdate field, execution on
AWS was trying to parse the empty string as a date.  Digging in logs in AWS
CloudWatch showed that the Unicode empty string, u'', was being passed to the
Lambda function.  Apparently, u'' isn't Falsy, so the existing "if start is not
None" check wasn't working as expected, so an additional "len(start)!=0" check
was added.